refactor: fix typo in TableCalculationModal name
